重力操作をstateパターンに変更しました。
