+++ Assets/Script/GravityController/AccelerationController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c4d4b18de1ab24bc79e5527008caca22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
