+++ Assets/Script/GravityController/JoystickController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ec868ef812e34b98bfe9fe373c4872e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
