+++ Assets/Script/SingletonMonoBehaviour.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SingletonMonoBehaviour<T> : MonoBehaviour where T :MonoBehaviour
+{
+    private static T instance;
+
+    public static T Instance
+    {
+        get
+        {
+            if (!instance)
+            {
+                instance = (T)FindObjectOfType(typeof(T));
+
+                if (instance == null)
+                {
+                    Debug.LogError(typeof(T) + "がシーンに存在しません。");
+                }
+            }
+            return instance;
+        }
+    }
+}
